refactor: 💡 (viva-ms) syncExpiryTime lambda
New code structure
